+++ src/Data/Barbie/Internal/Product.hs
@@ -5,7 +5,6 @@
 module Data.Barbie.Internal.Product
   ( ProductB(buniq, bprod)
   , bzip, bunzip, bzipWith, bzipWith3, bzipWith4
-  , (/*/), (/*)
 
   , CanDeriveProductB
   , GProductB(..)
@@ -16,7 +15,6 @@ where
 
 import Data.Barbie.Internal.Functor (FunctorB (..))
 
-import Data.Functor.Prod
 import Data.Functor.Product (Product (..))
 import Data.Kind            (Type)
 import Data.Proxy           (Proxy (..))
@@ -123,32 +121,6 @@ type CanDeriveProductB b f g
     , GProductB f g (RepN (b f)) (RepN (b g)) (RepN (b (f `Product` g)))
     )
 
-
--- | Like 'bprod', but returns a binary 'Prod', instead of 'Product', which
---   composes better.
---
---   See '/*/' for usage.
-(/*/)
-  :: ProductB b => b f -> b g -> b (Prod '[f, g])
-l /*/ r
-  = bmap (\(Pair f g) -> Cons f (Cons g Unit)) (l `bprod` r)
-infixr 4 /*/
-
--- | Similar to '/*/' but one of the sides is already a @'Prod' fs@.
---
---   Note that '/*', '/*/' and 'uncurryn' are meant to be used together:
---   '/*' and '/*/' combine @b f1, b f2...b fn@ into a single product that
---   can then be consumed by using `uncurryn` on an n-ary function. E.g.
---
--- @
--- f :: f a -> g a -> h a -> i a
---
--- 'bmap' ('uncurryn' f) (bf '/*' bg '/*/' bh)
--- @
-(/*) :: ProductB b => b f -> b (Prod fs) -> b (Prod (f ': fs))
-l /* r =
-  bmap (\(Pair f fs) -> oneTuple f `prod` fs) (l `bprod` r)
-infixr 4 /*
 
 -- ======================================
 -- Generic derivation of instances